app/backend/requirements/base.txt
Django==4.2.21
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
python-decouple==3.8
psycopg2-binary==2.9.9
